+++ frontend/components/course-degree-table/types.ts
@@ -1,0 +1,9 @@
+export type Item = {
+  id: string;
+  name: string;
+  email: string;
+  location: string;
+  flag: string;
+  status: "Active" | "Inactive" | "Pending";
+  balance: number;
+};
